Add a gray to white background gradient.
This was a pain to do: native CSS gradients shows banding. High and
narrow svgs with the gradient don't, but their repetition on the x-axis
is noticeable. I gravitated towards a grainy gradient, but the svg file
with the grainy gradient could not be rendered as-is in the browser.
Ended up creating a relatively narrow png (200x1200) with the gradient.
It works, but the file size is large (490kb). It'll do for now.